chore: update helper name & move from components/refactor to components
