+++ src/common/components/SubGroupInput.jsx
@@ -2,7 +2,7 @@ import { Fragment } from "react";
 import { useFieldArray, useFormContext } from "react-hook-form";
 import FormButtons from "./FormButtons";
 import FormInput from "./FormInput";
-import { getNestedError } from "./helper/getNestedError";
+import { getNestedError } from "../helpers/getNestedError";
 import Drag from "./Drag";
 import { Draggable } from "react-beautiful-dnd";
 import { useEffect } from "react";
