+++ src/common/helpers/turnGroupObject.jsx
