+++ src/modules/Portfolio.jsx
@@ -4,9 +4,9 @@ import useCusForm from "../common/hook/useCusForm";
 import Form from '../common/components/Form';
 import { useEffect } from "react";
 import { v4 as uuidv4 } from 'uuid';
-import { turnGroupObject } from "../common/components/helper/turnGroupObject";
+import { turnGroupObject } from "../common/helpers/turnGroupObject";
+import { turnArray } from "../common/helpers/turnArray";
 import SubGroupInput from "../common/components/SubGroupInput";
-import { turnArray } from "../common/components/helper/turnArray";
 import { useState } from "react";
 import PortfolioCard from "./PortfolioCard";
